requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langgraph
langchain-community
duckduckgo-search
wikipedia
streamlit
google-generativeai
sqlite3
pydantic
requests